requirements.txt
Flask~=1.1.2
python-dotenv
Flask-SQLAlchemy
pymysql